refactor: centralize the clippy policy and fix the surfaced violations
Move the clippy policy out of ci.yml's inline flags into
[workspace.lints.clippy] (deny clippy::all + the project's accepted allow-list,
plus result_large_err and approx_constant) so local `cargo clippy` enforces
exactly what CI does; CI's clippy step collapses to one line.

Enforcing the policy locally surfaced ~116 pre-existing clippy::all violations
(latent from the un-CI'd workspace restructuring). `cargo clippy --fix`
resolved the machine-applicable bulk; the rest are fixed by hand —
needless slice clones to std::slice::from_ref, needless ?Sized bounds,
items-after-test-module, collapsible if-let, doc-comment spacing, and
test-only &mut->&. No runtime behavior changes; the full clippy deny gate and
workspace check are clean and the consensus-crate test suites pass.

CONVENTIONS.md is rewritten with the full canonical standard (lints, errors,
manifests, macros, module layout, docs, naming, tests).

+++ neo-oracle-service/src/service/transactions/response.rs
@@ -170,7 +170,7 @@ impl OracleService {
         }
 
         size += get_var_size_serializable_slice(tx.attributes());
-        let comp3 = size as i64 * fee_per_byte as i64;
+        let comp3 = size as i64 * fee_per_byte;
         let final_network_fee = tx.network_fee().saturating_add(comp3);
         tx.set_network_fee(final_network_fee);
         tx.set_system_fee(request.gas_for_response - final_network_fee);
